backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
requests==2.31.0
gunicorn==21.2.0
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
lightfm==1.16
psycopg2-binary==2.9.1
alembic==1.7.1
httpx==0.19.0
